backend/requirements.txt
fastapi==0.99.1
uvicorn[standard]
SQLAlchemy==1.4.54
asyncpg
sqlalchemy[asyncio]
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
python-dotenv
pydantic==1.10.15
